+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## HEAD (Unreleased)
 
+### Improvements
+
+-   Optionally render YAML for k8s resources. (https://github.com/pulumi/pulumi-kubernetes/pull/936).
+
 ## 1.5.1 (February 7, 2020)
 
 ### Bug fixes
@@ -10,7 +14,6 @@
 
 ### Improvements
 
--   Optionally render YAML for k8s resources. (https://github.com/pulumi/pulumi-kubernetes/pull/936).
 -   Update nodejs SDK to use optional chaining in constructor. (https://github.com/pulumi/pulumi-kubernetes/pull/959).
 -   Automatically set Secret inputs as pulumi.secret. (https://github.com/pulumi/pulumi-kubernetes/pull/961).
 -   Create helm.v3 alias. (https://github.com/pulumi/pulumi-kubernetes/pull/970).